refactored faq into new file to distinguish between faq and branchfaq
