+++ src/app/page.tsx
@@ -4,7 +4,7 @@ import AboutBlurb from "../components/landing/body/AboutBlurb";
 import Branches from "../components/landing/body/BranchSection/Branches";
 import EventsSection from "../components/landing/body/EventsSection/EventsSection";
 import Footer from "../components/shared/Footer/Footer";
-import FAQ from "@/components/landing/body/FAQ";
+import FAQ from "@/components/landing/body/FAQ/FAQ";
 
 export default function Home() {
   return (
